requirements.txt
psycopg2-binary
python-dotenv
sqlalchemy
streamlit==1.37.0

